+++ src/client/Microsoft.Identity.Client/OAuth2/OAuthConstants.cs
@@ -66,7 +66,6 @@ namespace Microsoft.Identity.Client.OAuth2
     internal static class OAuth2AssertionType
     {
         public const string JwtBearer = "urn:ietf:params:oauth:client-assertion-type:jwt-bearer";
-        public const string FmiBearer = "urn:ietf:params:oauth:client-assertion-type:fmi-bearer";
     }
 
     internal static class OAuth2RequestedTokenUse